Errors cleared

+++ helpers/cart/cart.ts
@@ -1,23 +1,29 @@
 // Types
 import { Dispatch, SetStateAction } from "react"
 import { CartType } from "../../types/Cart"
+import { UserType } from "../../types/User"
 
-export const AddToCart = (uid: number, pid: number, quantity: number, setCart: Dispatch<SetStateAction<CartType>>, image: string, price: number, name: string, setIsCartOpen: (isCartOpen: boolean) => void) => new Promise<{id: number}>(async(resolve, reject) => {
-    try {
-        const response = await fetch(`/api/cart`, {
-            method: "POST",
-            headers:{
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify({
-                uid: uid,
-                pid: pid,
-                quantity: quantity
+export const AddToCart = (user: UserType | null, pid: number, quantity: number, setCart: Dispatch<SetStateAction<CartType>>, image: string, price: number, name: string, setIsCartOpen: (isCartOpen: boolean) => void) => new Promise<{id: number}>(async(resolve, reject) => {
+    if(user !== null) {
+        const uid = user.id;
+        try {
+            const response = await fetch(`/api/cart`, {
+                method: "POST",
+                headers:{
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify({
+                    uid: uid,
+                    pid: pid,
+                    quantity: quantity
+                })
             })
-        })
-        resolve(response.json())
-    } catch (err) {
-        reject(err)
+            resolve(response.json())
+        } catch (err) {
+            reject(err)
+        }
+    } else {
+        alert("You must be logged in to add items to your cart!")
     }
 }).then((newItem) => {
     setCart(old => {
@@ -53,24 +59,29 @@ export const AddToCart = (uid: number, pid: number, quantity: number, setCart: D
     setIsCartOpen(true)
 })// TODO
 
-export const UpdateCartItemQuantity = (uid: number, id: number, quantity: number, type: string, setCart: Dispatch<SetStateAction<CartType>>) => new Promise(async(resolve, reject) => {
+export const UpdateCartItemQuantity = (user: UserType | null, id: number, quantity: number, type: string, setCart: Dispatch<SetStateAction<CartType>>) => new Promise(async(resolve, reject) => {
     if(quantity == 1 && type == "decrement") return;
     else {
-        try {
-            const response = await fetch(`/api/cart`, {
-                method: "PUT",
-                headers:{
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify({
-                    id: id,
-                    uid: uid,
-                    type: type
+        if(user !== null) {
+            const uid = user.id;
+            try {
+                const response = await fetch(`/api/cart`, {
+                    method: "PUT",
+                    headers:{
+                        "Content-Type": "application/json"
+                    },
+                    body: JSON.stringify({
+                        id: id,
+                        uid: uid,
+                        type: type
+                    })
                 })
-            })
-            resolve(response.json())
-        } catch (err) {
-            reject(err)
+                resolve(response.json())
+            } catch (err) {
+                reject(err)
+            }
+        } else {
+            alert("You must be logged in to update your cart")
         }
     }
 
@@ -78,21 +89,26 @@ export const UpdateCartItemQuantity = (uid: number, id: number, quantity: number
     {...old, items: old.items.map(item => item.id === id ? {...item, quantity: type == "increment" ? item.quantity + 1 : item.quantity - 1} : item)}
 )))
 
-export const RemoveItemFromCart = (uid: number, id: number, setCart: Dispatch<SetStateAction<CartType>>) => new Promise(async(resolve, reject) => {
-    try {
-        const response = await fetch(`/api/cart`, {
-            method: "DELETE",
-            headers:{
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify({
-                id: id,
-                uid: uid
+export const RemoveItemFromCart = (user: UserType | null, id: number, setCart: Dispatch<SetStateAction<CartType>>) => new Promise(async(resolve, reject) => {
+    if(user !== null) {
+        const uid = user.id;
+        try {
+            const response = await fetch(`/api/cart`, {
+                method: "DELETE",
+                headers:{
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify({
+                    id: id,
+                    uid: uid
+                })
             })
-        })
-        resolve(response.json())
-    } catch (err){
-        reject(err)
+            resolve(response.json())
+        } catch (err){
+            reject(err)
+        }
+    } else {
+        alert("You must be logged in to remove items from your cart")
     }
 }).then(() => setCart(old => (
     {...old, items: old.items.filter(item => item.id !== id)}
